// generated/src/app/api-models/inst-claim-oth-carr-edi-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class InstClaimOthCarrEdiW {
  seqPrediId: number;
  seqClaimId: number;
  claimNumber: string;
  originalClaimNumber: string;
  sequenceNo: string;
  payerName: string;
  providerNo: string;
  patReleaseInd: string;
  assignBenefitsInd: string;
  priorPayment: string;
  estAmountDue: string;
  insuredLastName: string;
  insuredFirstName: string;
  insuredMiddleName: string;
  patRelToInsured: string;
  idNo: string;
  groupName: string;
  groupNo: string;
  treatmentAuthCode: string;
  employmentStatusCode: string;
  insuredEmpName: string;
  insuredEmpLocation: string;
  internalControlNo: string;
  ediStatus: string;
  recInd: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}